Fix #3: utilities separate file with definitions and prototypes
