+++ elgamal.h
@@ -4,6 +4,7 @@
  */
 
 #include "utilities.h"
+#include <math.h>
 
 /**
  * Struct with functions to generate keys and
